Don't understand... give up the compilation mode.

+++ src/execution/sema/sema_builtin.cpp
@@ -855,7 +855,7 @@ void Sema::CheckBuiltinExecutionContextCall(ast::CallExpr *call, UNUSED_ATTRIBUT
     case ast::Builtin::ExecutionContextStartResourceTracker: {
       // MetricsComponent
       if (!call_args[1]->IsIntegerLiteral()) {
-        ReportIncorrectCallArg(call, 1, GetBuiltinType(ast::BuiltinType::Uint8));
+        ReportIncorrectCallArg(call, 1, GetBuiltinType(ast::BuiltinType::Uint64));
         return;
       }
       // Init returns nil
